+++ admin-authority-sample-embedded/tests/ui/missing_admin_initialize.rs
@@ -1,5 +1,5 @@
-// A field before the slot changed without the marker offset following.
-// The emitted agreement assert must refuse this build.
+// The account-creating instruction lacks #[admin_initialize]. The
+// framework must refuse a program that ships born renounced.
 use admin_authority::{AdminCandidate, AdminConfig};
 use borsh::{BorshDeserialize, BorshSerialize};
 use spel_framework::prelude::*;
@@ -7,7 +7,7 @@ use spel_framework::prelude::*;
 #[account_type]
 #[derive(BorshSerialize, BorshDeserialize, Clone, Debug, Default)]
 pub struct ProgConfig {
-    pub value: u32,
+    pub value: u64,
     pub padding: [u8; 24],
     #[admin_slot]
     pub admin: AdminConfig,
@@ -16,10 +16,9 @@ pub struct ProgConfig {
 #[lez_program]
 #[admin_authority(admin_config = config, offset = 32)]
 mod fixture {
-    use admin_authority::{admin_initialize, require_admin};
+    use admin_authority::require_admin;
 
     #[instruction]
-    #[admin_initialize]
     pub fn initialize(
         #[account(init, pda = literal("prog_config"))] mut config: AccountWithMetadata,
         #[account(signer)] signer: AccountWithMetadata,
